// src/app/shared/models/vehicles.model.ts
export interface Vehicle {
    id: number,
    createdBy: string,
    plate: string;
    chassisNumber: string;
    transitLicense: string;
    engineNumber: string;
    vin: string;
    cylinderCapacity:number;
    model: number;
    emptyWeight: number;
    loadingCapacity: number;
    fuelType: string;
    propertyType: string;
    vehicleConfiguration: string;
    brand: string;
    vehicleLine: string;
    bodyWork: string;
    company: string;
    status: string
}